requirements.txt
Django~=2.2.2
djangorestframework~=3.9.4
django-filter~=2.1.0
psycopg2-binary~=2.8.3
djangorestframework-api-key~=1.2.1